+++ rust/operator-binary/src/security/opa.rs
@@ -19,6 +19,7 @@ pub enum Error {
 
 type Result<T, E = Error> = std::result::Result<T, E>;
 
+#[derive(Clone, Debug)]
 pub struct HbaseOpaConfig {
     authorization_connection_string: String,
     dry_run: bool,